Move ConfigurationChangeModel to plugin
Contributes to CL-897
